# Bit-exact version of Luv2RGB_b