### This pullrequest changes

1. Luv2RGB_b is now calculated in integers which provides bit-exactness across architectures, compilers, etc. All lookup tables are calculated in float
2. OCL code is done in correspondence with CPU's, docs fixed about X, Y, Z range
3. Added bit-exactness test for Luv2RGB_b

### Performance

Measured 100 runs on random 257x257 image
Repeated 100 times and then averaged

Version | Linux x64 Core i5-6600 @ 3.3 GHz |  Win x64 Core i5-6300U @ 2.4 GHz
------------ | ------------- | -------------
Bit-exact | 0.091 s | 0.195 s
Original | 0.051 s | 0.083 s

### Accuracy

Range of values is 0 to 255

- max abs error: 2 **ONLY FOR VALUES WHICH LIE IN WHITE POINT RANGE (WHEN 0 <= X, Y, Z < 2)**
- mean by RGB channels: mean [0.02, 0.01, 0.008]
- stddev by RGB channels: [0.13, 0.12, 0.09]