@savuor 2017-09-13
Perf data updated by [PR #376](https://github.com/opencv/opencv_extra/pull/376) in extra repo